requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot==21.5
apscheduler==3.10.4
httpx==0.27.0
